TUSS4470_shield_002/requirements.txt
numpy
pyserial==3.4
PyQt5
pyqtgraph
pyqtdarktheme
